updated tests for SingleDartTest and DartTowerTest

+++ src/ooga/backend/darts/SingleDart.java
@@ -4,7 +4,7 @@ import ooga.backend.API.Moveable;
 
 public class SingleDart extends Dart {
 
-  public SingleDart(int xPosition, int yPosition, int xVelocity, int yVelocity) {
+  public SingleDart(int xPosition, int yPosition, double xVelocity, double yVelocity) {
     super(xPosition, yPosition, xVelocity, yVelocity);
   }
 }
